指令的基本执行

+++ main02/rtda/local_var.go
@@ -5,7 +5,7 @@ import "math"
 //slot的数组
 type LocalVars []Slot
 
-func newLocalVars(maxLocals uint) LocalVars {
+func newLocalVars(maxLocals uint16) LocalVars {
 
 	if maxLocals > 0 {
 		return make([]Slot, maxLocals)
